requirements.txt
discord.py
python-dotenv
langchain
langchain-community
langchain-core
chromadb
sentence-transformers
faiss-cpu
numpy
PyYAML
pytest
pytest-asyncio
